test: split governance vote tally fixtures
Replace the monolithic GovernanceVoteTallyIT with fixture-scoped classes for DRep, DRep lifecycle, SPO, and committee rows. The split keeps shared setup within compatible rule families while avoiding state leakage from ratified constitutions, DRep deregistration, and committee mutations. Update the test inventory and keep the shared DevKit setup in AbstractGovernanceVoteTallyIT.
